requirements.txt
PyYAML==6.0.2
requests==2.32.3
pyspark==3.5.5
matplotlib==3.10.1
pyiceberg==0.9.0